int voor budget naar double veranderd, in de listMovieDisplay meer data meegegeven (heeft error met deleten, kijk in model/controller om het te fixen).

+++ src/eindproject/AddMovieDisplay.java
@@ -63,7 +63,7 @@ public class AddMovieDisplay extends View {
         return (String) comboBox1.getSelectedItem();
     }
 
-    private int getBudget() {
+    private double getBudget() {
         return Integer.parseInt(textField4.getText());
     }
 }
